fix resolving .lnk links in fs::EnumFiles();  pass fs::CFileState instead of CFileFind in fs::IEnumerator interface to minimize data conversions during file enumeration;  extract found file uniqueness in fs::CBaseEnumerator class.

+++ cmd_line/TransferFiles/XferOptions.h
@@ -48,8 +48,8 @@ public:
 	std::vector< std::tstring > m_excludeFindSpecs;
 	std::auto_ptr< fs::CPath > m_pBackupDirPath;
 
-	DWORD m_mustHaveFileAttr;
-	DWORD m_mustNotHaveFileAttr;
+	BYTE m_mustHaveFileAttr;
+	BYTE m_mustNotHaveFileAttr;
 
 	CheckFileChanges m_filterBy;
 	CTime m_earliestTimestamp;
